tests|github|pre-commit: Improve `build-workflows.py` and make it a `pre-commit` hook (#8728)
* tests: Move `default_replacements` into `generate_replacements`

* tests: Directly use `root_test_dirs`

Instead of passing it into `subdirs`

* tests: Make `build-workflow.py` callable from everywhere

* tests: Adjust the output of `build-workflows.py`

* tests: Add `-f` to `build-workflows.py` to allow failing on CI

* github: Run `build-workflows.py`

* pre-commit: Run `build-workflows.py` as first hook

* tests: Skip DID wallet tests until they are reliable and fixed

* tests: Skip `test_using_legacy_cryptfilekeyring`

* tests|github: Drop `root_test_dirs` and just search in all directories

This removes the requirement to add new test directories to the 
`testconfig.py`. This change uncovered two more tests which were not 
enabled on CI because their parent dir's were not added to 
`root_test_dirs`.
- `tests/weight_proof/test_weight_proof.py`
- `tests/util/test_struct_stream.py`

* tests|github: Don't list test files, use `tests/dir/tests_*.py` instead

* test: Improve workflow update checks by comparing old vs new

Instead of doing a `git diff` which can lead to false positives if there
are other local changes unlrelated to test workflows. This also fixes
the issue that newly added files were not tracked by `git diff` which
means the script / the CI would trigger an alert.

* tests: Adjust `build-workflows.py` to work as expected on windows

* tests: Read/Write in bytes to avoid and detect unexpected `\r\n`

* tests: Skip `test_pool_cmdline.py`